Generate cross-link and external links
This finishes up the previous commit
by generating cross-package links
for entities referenced in godoc comments
as well as entities referenced in generated declarations.

The '-link' flag was renamed to '-pkg-doc' to be more accurate,
as it only generates links to package documentation.
We internally generate the entity fragment links.

This required a some restructuring
because we need the full package list
to know which packages we can generate relative references to.

Resolves #2