template: Add missing method anchors
Methods were missing the `$T.$M` anchor
because they were relying on the function template
which only sets up anchors based on the function name.

Resolving this requires some duplication because
the functions/methods for types are h4
but regular functions are h3.
It's possible to DRY this with more template complexity,
but I think that's not worth it.

Resolves #11