gosrc/finder: Support pkg name != basename path
To handle the case where the base name of an import path
does not match the name of the package inside it,
(e.g. example.com/foo-go would typically have 'foo')
and accurately map 'foo' in 'foo.Bar' to the foo-go import path,
we need to specify an ast.Importer to ast.NewPackage.

We can use the NeedImports option of go/packages
to get information about packages imported by a package
and use that to build an accurate ast.Importer.