html: Renderer should know about -internal
Generator currently inspects the -internal flag to decide
which subpackages to feed to the Renderer
for a package or directory's subpackage listing.

It makes better sense for this information to be made
available to the Renderer and have it decide
whether it wants to render internal packages or not.

This change moves the Internal bool to the Renderer.
This also helped DRY up the tests for subpackage listings
and breadcrumbs.