Add -frontmatter flag
Adds a -frontmatter and -frontmatter-file flag
that lets users specify a text/template or a file containing one.

If specified, this template will be run for each generated HTML file,
and the result will be prepended to the top of each HTML file,
separated from the rest of the content with an empty line.

Although text/template is a bit rough,
this provides enough flexibility to use doc2go with a variety
of different Hugo and Jekyll templates.

Resolves #14