Add syntax highlighting
Add support for syntax highlighting in generated declarations.
This is accomplished with the help of the Chroma syntax highlighter.
The output of the highlighter, and our own labels are interepersed,
with our labels taking precedence.

Adds the following new flags:

    -highlight [MODE:]THEME
    -highlight-print-css
    -highlight-list-themes

MODE can be 'inline' for inline `style` attributes,
'classes' for `class` attributes and an accompanying stylesheet,
or 'auto' to pick based on whether we're in `-embed` mode or not.

Adds a 'plain' theme that replicates the old highlighting.
It leaves everything except comments as regular text,
and comments get a slight fade.

Resolves #62