Respect -rel-link-style=directory for code links
When we added support for -rel-link-style,
only links generated by the HTML package directly were updated.
We missed links generated inside code samples.

This fixes those links by plumbing RelLinkStyle to docLinker,
and using it to normalize them.

Adds an integration test to verify that all relative links
in the generated HTML have a trailing '/' with this flag
and can still be visited.

Resolves #154