standalone/navbar: Add "Root" link
Adds a link that goes to the root of the website to the navbar.
This link will point to the index.html inside OutDir, whatever that is.
By default, that's the top-level package listing.
With `-subdir`, that's the sibling directory listing.

In adding this, I discovered and fixed a bug
where an "Index" link would show even on directory listing pages
that had breadcrumbs.

Resolves #187