Add client-side search with Pagefind
Adds client-side using [Pagefind](https://pagefind.app).
To configure this functionality, there are three options:

- By default, if a pagefind binary is found on $PATH,
  the feature is enabled.
- It can be explicitly enabled by passing --pagefind
  or --paegfind=path/to/binary with an explicit path to pagefind.
- It can be explicitly disabled by passing --pagefind=false.

This feature cannot be used in combination with --embed.

To make this nice, break RenderSiteIndex off from RenderPackageIndex.
The root level site for -subdir is different from a package index.
It has a different template, and does not get a pagefind search box.

Resolves #75