Don't render types and functions in main packages
For main packages, the only consumable part is the command.
There's no need to include the documentation of exported types defined
in main packages, because it's impossible to import and consume them.

To do this, a new command.html template is introduced.
This allows removing command-specific special cases from the package.html.

Resolves #210