highlight: fix index panic on certain inputs
It's possible for TokenIndex.Interval to panic
if we're trying to capture trail text for a token
(because the end offset is in the middle of a token)
but the start of that token is at or after the requested end offset.
This will usually happen with chaotic whitespace.

The included unit test reproduces the panic without the fix.

Resolves #290